hwcontext: add an AVHWFramesInternal.internal_ref field
The issue is that Vulkan requires an extra structure, whose pointer
is set in hwfc->create_pnext. That structure is required for video
decoding and must always exist for the entire duration of the frames
context.

The original idea of trying to use the decoder context was a failure.
The decoder context lives for a shorter amount of time than the frames
context. During seeking/flushing, in case of H264, a new context is
immediately created, with the old one destroyed.
However, the frames context survived, and when requested by the decoder
to allocate a frame, it failed, as the structure had already been
deallocated, leaving a dangling pointer in create_pnext.

Instead of this mess, create a new AVHWFramesInternal.internal_ref field,
which is unreferenced during context destruction.
Due to the way that video decoding works, in that the frames context
is always created by libavcodec, we're free to do this and set our
own pointer for the frames context we create.